Split asm using autodecompiler

+++ include/common_structs.h
@@ -49,8 +49,8 @@ typedef struct {
 } StaffGhost;
 // 80160ADC Banana's remaining
 typedef struct {
-    s32 unk0;
-    s32 unk2;
+    s16 unk0;
+    s16 unk2;
     s32 shellPos1; // BF8 = empty
     s32 shellPos2; // 422 full
     s32 shellPos3; // 422C full
